Create CrawlCrunchbaseInvestors.js.
This is designed to run directly on browser console, after injected jQuery.

copy this page entirely into the console, and run. It will get the data on the table on the page, and reload to next page for next page of data, til the end.

At the end, there will be data in the variable allProcessedData. you can save it.